TurnosxEspecialidades
Ahora cada turno tiene su respectiva especialidad, se carga correctamente en la aplicación, y se muestran turnos en base a la especialidad al modificar en DetalleTurno

+++ Proyecto-Clinica/Conexion-Clinica/EspecialidadesConexion.cs
@@ -25,7 +25,7 @@ namespace Conexion_Clinica
                 {
                     Especialidad especialidad = new Especialidad
                     {
-                        Id = (byte)datos.Lector["ID_ESPECIALIDAD"],
+                        Id = (int)datos.Lector["ID_ESPECIALIDAD"],
                         Tipo = (string)datos.Lector["TIPO"]
                     };
                     lista.Add(especialidad);
